fix: element.offsetTop  🔽
offsetTop only gets offset relative to the PARTENT element. In our testing so far that happened to be the entire body, so it wasn't caught yet. The simple solution is to get the element's boundingClientRect() object, and top from that (this will always be relative to the part of the page you can view. Add window.scrollY to that value and you get a reliable "absolutely Y position on the page" for the element, give or take .0001px 😂